Replication:refactor for better separation of concerns and documentation
