+++ usecases/replica/finder.go
@@ -15,8 +15,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sort"
-	"strings"
 
 	"github.com/go-openapi/strfmt"
 	"github.com/sirupsen/logrus"
@@ -24,29 +22,19 @@ import (
 	"github.com/weaviate/weaviate/entities/search"
 	"github.com/weaviate/weaviate/entities/storobj"
 	"github.com/weaviate/weaviate/usecases/objects"
-	"golang.org/x/sync/errgroup"
 )
 
 var (
 	// msgCLevel consistency level cannot be achieved
 	msgCLevel = "cannot achieve consistency level"
-	// errConflictFindDeleted object exists on one replica but is deleted on the other.
-	//
-	// It depends on the order of operations
-	// Created -> Deleted    => It is safe in this case to propagate deletion to all replicas
-	// Created -> Deleted -> Created => propagating deletion will result in data lost
-	errConflictExistOrDeleted = errors.New("conflict: object has been deleted on another replica")
 
-	// errConflictObjectChanged object changed since last time and cannot be repaired
-	errConflictObjectChanged = errors.New("source object changed during repair")
-
-	errReplicas = errors.New("cannot find enough replicas")
+	errReplicas = errors.New("cannot reach enough replicas")
 	errRepair   = errors.New("read repair error")
 	errRead     = errors.New("read error")
 )
 
 type (
-	// senderReply represent the data received from a sender
+	// senderReply is a container for the data received from a replica
 	senderReply[T any] struct {
 		sender     string // hostname of the sender
 		Version    int64  // sender's current version of the object
@@ -63,35 +51,40 @@ type (
 
 // Finder finds replicated objects
 type Finder struct {
-	client   finderClient // needed to commit and abort operation
-	resolver *resolver    // host names of replicas
-	class    string
-	// TODO LOGGER
-	// Don't leak host nodes to end user but log it
-	log logrus.FieldLogger
+	resolver     *resolver // host names of replicas
+	finderStream           // stream of objects
 }
 
 // NewFinder constructs a new finder instance
 func NewFinder(className string,
-	stateGetter shardingState, nodeResolver nodeResolver,
-	client RClient, l logrus.FieldLogger,
+	stateGetter shardingState,
+	nodeResolver nodeResolver,
+	client rClient,
+	l logrus.FieldLogger,
 ) *Finder {
+	cl := finderClient{client}
 	return &Finder{
-		client: finderClient{client},
 		resolver: &resolver{
 			schema:       stateGetter,
 			nodeResolver: nodeResolver,
 			class:        className,
 		},
-		class: className,
-		log:   l,
+		finderStream: finderStream{
+			repairer: repairer{
+				class:  className,
+				client: cl,
+			},
+			log: l,
+		},
 	}
 }
 
 // GetOne gets object which satisfies the giving consistency
 func (f *Finder) GetOne(ctx context.Context,
 	l ConsistencyLevel, shard string,
-	id strfmt.UUID, props search.SelectProperties, adds additional.Properties,
+	id strfmt.UUID,
+	props search.SelectProperties,
+	adds additional.Properties,
 ) (*storobj.Object, error) {
 	c := newReadCoordinator[findOneReply](f, shard)
 	op := func(ctx context.Context, host string, fullRead bool) (findOneReply, error) {
@@ -114,27 +107,16 @@ func (f *Finder) GetOne(ctx context.Context,
 		return nil, fmt.Errorf("%s %q: %w", msgCLevel, l, errReplicas)
 	}
 	result := <-f.readOne(ctx, shard, id, replyCh, state)
-	if err = result.err; err != nil {
+	if err = result.Err; err != nil {
 		err = fmt.Errorf("%s %q: %w", msgCLevel, l, err)
 	}
-	return result.data, err
-}
-
-// batchReply represents the data returned by sender
-// The returned data may result from a full or digest read request
-type batchReply struct {
-	// Sender hostname of the sender
-	Sender string
-	// IsDigest is this reply from a digest read?
-	IsDigest bool
-	// FullData returned from a full read request
-	FullData []objects.Replica
-	// DigestData returned from a digest read request
-	DigestData []RepairResponse
+	return result.Value, err
 }
 
 // GetAll gets all objects which satisfy the giving consistency
-func (f *Finder) GetAll(ctx context.Context, l ConsistencyLevel, shard string,
+func (f *Finder) GetAll(ctx context.Context,
+	l ConsistencyLevel,
+	shard string,
 	ids []strfmt.UUID,
 ) ([]*storobj.Object, error) {
 	c := newReadCoordinator[batchReply](f, shard)
@@ -153,15 +135,19 @@ func (f *Finder) GetAll(ctx context.Context, l ConsistencyLevel, shard string,
 		return nil, fmt.Errorf("%s %q: %w", msgCLevel, l, errReplicas)
 	}
 	result := <-f.readAll(ctx, shard, ids, replyCh, state)
-	if err = result.err; err != nil {
+	if err = result.Err; err != nil {
 		err = fmt.Errorf("%s %q: %w", msgCLevel, l, err)
 	}
 
-	return result.data, err
+	return result.Value, err
 }
 
 // Exists checks if an object exists which satisfies the giving consistency
-func (f *Finder) Exists(ctx context.Context, l ConsistencyLevel, shard string, id strfmt.UUID) (bool, error) {
+func (f *Finder) Exists(ctx context.Context,
+	l ConsistencyLevel,
+	shard string,
+	id strfmt.UUID,
+) (bool, error) {
 	c := newReadCoordinator[existReply](f, shard)
 	op := func(ctx context.Context, host string, _ bool) (existReply, error) {
 		xs, err := f.client.DigestReads(ctx, host, f.class, shard, []strfmt.UUID{id})
@@ -177,16 +163,18 @@ func (f *Finder) Exists(ctx context.Context, l ConsistencyLevel, shard string, i
 		return false, fmt.Errorf("%s %q: %w", msgCLevel, l, errReplicas)
 	}
 	result := <-f.readExistence(ctx, shard, id, replyCh, state)
-	if err = result.err; err != nil {
+	if err = result.Err; err != nil {
 		err = fmt.Errorf("%s %q: %w", msgCLevel, l, err)
 	}
-	return result.data, err
+	return result.Value, err
 }
 
 // NodeObject gets object from a specific node.
 // it is used mainly for debugging purposes
 func (f *Finder) NodeObject(ctx context.Context,
-	nodeName, shard string, id strfmt.UUID,
+	nodeName,
+	shard string,
+	id strfmt.UUID,
 	props search.SelectProperties, adds additional.Properties,
 ) (*storobj.Object, error) {
 	host, ok := f.resolver.NodeHostname(nodeName)
@@ -195,430 +183,4 @@ func (f *Finder) NodeObject(ctx context.Context,
 	}
 	r, err := f.client.FullRead(ctx, host, f.class, shard, id, props, adds)
 	return r.Object, err
-}
-
-func (f *Finder) readOne(ctx context.Context,
-	shard string,
-	id strfmt.UUID,
-	ch <-chan simpleResult[findOneReply],
-	st rState,
-) <-chan result[*storobj.Object] {
-	// counters tracks the number of votes for each participant
-	resultCh := make(chan result[*storobj.Object], 1)
-	go func() {
-		defer close(resultCh)
-		var (
-			votes      = make([]objTuple, 0, len(st.Hosts))
-			maxCount   = 0
-			contentIdx = -1
-		)
-
-		for r := range ch { // len(ch) == st.Level
-			resp := r.Response
-			if r.Err != nil { // a least one node is not responding
-				f.log.WithField("op", "get").WithField("replica", resp.sender).
-					WithField("class", f.class).WithField("shard", shard).WithField("uuid", id).Error(r.Err)
-				resultCh <- result[*storobj.Object]{nil, errRead}
-				return
-			}
-			if !resp.DigestRead {
-				contentIdx = len(votes)
-			}
-			votes = append(votes, objTuple{resp.sender, resp.UpdateTime, resp.Data, 0, nil})
-			for i := range votes { // count number of votes
-				if votes[i].UTime == resp.UpdateTime {
-					votes[i].ack++
-				}
-				if maxCount < votes[i].ack {
-					maxCount = votes[i].ack
-				}
-				if maxCount >= st.Level && contentIdx >= 0 {
-					resultCh <- result[*storobj.Object]{votes[contentIdx].o.Object, nil}
-					return
-				}
-			}
-		}
-
-		obj, err := f.repairOne(ctx, shard, id, votes, st, contentIdx)
-		if err == nil {
-			resultCh <- result[*storobj.Object]{obj, nil}
-			return
-		}
-
-		resultCh <- result[*storobj.Object]{nil, errRepair}
-		var sb strings.Builder
-		for i, c := range votes {
-			if i != 0 {
-				sb.WriteByte(' ')
-			}
-			fmt.Fprintf(&sb, "%s:%d", c.sender, c.UTime)
-		}
-		f.log.WithField("op", "repair_one").WithField("class", f.class).WithField("shard", shard).WithField("uuid", id).WithField("msg", sb.String()).Error(err)
-	}()
-	return resultCh
-}
-
-// repair one object on several nodes using last write wins strategy
-func (f *Finder) repairOne(ctx context.Context, shard string, id strfmt.UUID, votes []objTuple, st rState, contentIdx int) (_ *storobj.Object, err error) {
-	var (
-		lastUTime int64
-		winnerIdx int
-		cl        = f.client
-	)
-	for i, x := range votes {
-		if x.o.Deleted {
-			return nil, errConflictExistOrDeleted
-		}
-		if x.UTime > lastUTime {
-			lastUTime = x.UTime
-			winnerIdx = i
-		}
-	}
-	// fetch most recent object
-	updates := votes[contentIdx].o
-	winner := votes[winnerIdx]
-	if updates.UpdateTime() != lastUTime {
-		updates, err = cl.FullRead(ctx, winner.sender, f.class, shard, id,
-			search.SelectProperties{}, additional.Properties{})
-		if err != nil {
-			return nil, fmt.Errorf("get most recent object from %s: %w", winner.sender, err)
-		}
-		if updates.UpdateTime() != lastUTime {
-			return nil, fmt.Errorf("fetch new state from %s: %w, %v", winner.sender, errConflictObjectChanged, err)
-		}
-	}
-
-	var gr errgroup.Group
-	for _, vote := range votes { // repair
-		if vote.UTime == lastUTime {
-			continue
-		}
-		vote := vote
-		gr.Go(func() error {
-			ups := []*objects.VObject{{
-				LatestObject:    &updates.Object.Object,
-				StaleUpdateTime: vote.UTime,
-			}}
-			resp, err := cl.Overwrite(ctx, vote.sender, f.class, shard, ups)
-			if err != nil {
-				return fmt.Errorf("node %q could not repair object: %w", vote.sender, err)
-			}
-			if len(resp) > 0 && resp[0].Err != "" {
-				return fmt.Errorf("overwrite %w %s: %s", errConflictObjectChanged, vote.sender, resp[0].Err)
-			}
-			return nil
-		})
-	}
-
-	return updates.Object, gr.Wait()
-}
-
-type vote struct {
-	batchReply
-	Count []int
-	Err   error
-}
-
-func (r batchReply) UpdateTimeAt(idx int) int64 {
-	if len(r.DigestData) != 0 {
-		return r.DigestData[idx].UpdateTime
-	}
-	return r.FullData[idx].UpdateTime()
-}
-
-type _Results result[[]*storobj.Object]
-
-func (f *Finder) readAll(ctx context.Context, shard string, ids []strfmt.UUID, ch <-chan simpleResult[batchReply], st rState) <-chan _Results {
-	resultCh := make(chan _Results, 1)
-
-	go func() {
-		defer close(resultCh)
-		var (
-			N = len(ids) // number of requested objects
-			// votes counts number of votes per object for each node
-			votes      = make([]vote, 0, len(st.Hosts))
-			contentIdx = -1 // index of full read reply
-		)
-
-		for r := range ch { // len(ch) == st.Level
-			resp := r.Response
-			if r.Err != nil { // a least one node is not responding
-				f.log.WithField("op", "get").WithField("replica", r.Response.Sender).
-					WithField("class", f.class).WithField("shard", shard).Error(r.Err)
-				resultCh <- _Results{nil, errRead}
-				return
-			}
-			if !resp.IsDigest {
-				contentIdx = len(votes)
-			}
-
-			votes = append(votes, vote{resp, make([]int, N), nil})
-			M := 0
-			for i := 0; i < N; i++ {
-				max := 0
-				lastTime := resp.UpdateTimeAt(i)
-
-				for j := range votes { // count votes
-					if votes[j].UpdateTimeAt(i) == lastTime {
-						votes[j].Count[i]++
-					}
-					if max < votes[j].Count[i] {
-						max = votes[j].Count[i]
-					}
-				}
-				if max >= st.Level {
-					M++
-				}
-			}
-
-			if M == N {
-				resultCh <- _Results{fromReplicas(votes[contentIdx].FullData), nil}
-				return
-			}
-		}
-		res, err := f.repairAll(ctx, shard, ids, votes, st, contentIdx)
-		if err == nil {
-			resultCh <- _Results{res, nil}
-		}
-		resultCh <- _Results{nil, errRepair}
-		f.log.WithField("op", "repair_all").WithField("class", f.class).
-			WithField("shard", shard).WithField("uuids", ids).Error(err)
-	}()
-
-	return resultCh
-}
-
-func (f *Finder) repairAll(ctx context.Context,
-	shard string,
-	ids []strfmt.UUID,
-	votes []vote,
-	st rState,
-	contentIdx int,
-) ([]*storobj.Object, error) {
-	var (
-		result     = make([]*storobj.Object, len(ids)) // final result
-		lastTimes  = make([]iTuple, len(ids))          // most recent times
-		ms         = make([]iTuple, 0, len(ids))       // mismatches
-		nDeletions = 0
-		cl         = f.client
-	)
-	// find most recent objects
-	for i, x := range votes[contentIdx].FullData {
-		lastTimes[i] = iTuple{S: contentIdx, O: i, T: x.UpdateTime(), Deleted: x.Deleted}
-	}
-	for i, vote := range votes {
-		if i != contentIdx {
-			for j, x := range vote.DigestData {
-				deleted := lastTimes[j].Deleted || x.Deleted
-				if x.UpdateTime > lastTimes[j].T {
-					lastTimes[j] = iTuple{S: i, O: j, T: x.UpdateTime}
-				}
-				lastTimes[j].Deleted = deleted
-			}
-		}
-	}
-	// find missing content (diff)
-	for i, p := range votes[contentIdx].FullData {
-		if lastTimes[i].Deleted { // conflict
-			nDeletions++
-			result[i] = nil
-		} else if contentIdx != lastTimes[i].S {
-			ms = append(ms, lastTimes[i])
-		} else {
-			result[i] = p.Object
-		}
-	}
-	if len(ms) > 0 { // fetch most recent objects
-		// partition by hostname
-		sort.SliceStable(ms, func(i, j int) bool { return ms[i].S < ms[j].S })
-		partitions := make([]int, 0, len(votes)-2)
-		pre := ms[0].S
-		for i, y := range ms {
-			if y.S != pre {
-				partitions = append(partitions, i)
-				pre = y.S
-			}
-		}
-		partitions = append(partitions, len(ms))
-
-		// concurrent fetches
-		gr, ctx := errgroup.WithContext(ctx)
-		start := 0
-		for _, end := range partitions { // fetch diffs
-			receiver := votes[ms[start].S].Sender
-			query := make([]strfmt.UUID, end-start)
-			for j := 0; start < end; start++ {
-				query[j] = ids[ms[start].O]
-				j++
-			}
-			start := start
-			gr.Go(func() error {
-				resp, err := cl.FullReads(ctx, receiver, f.class, shard, query)
-				if err != nil {
-					return err
-				}
-				for i, n := 0, len(query); i < n; i++ {
-					idx := ms[start-n+i].O
-					if lastTimes[idx].T != resp[i].UpdateTime() {
-						return fmt.Errorf("object %s changed on %s", ids[idx], receiver)
-					}
-					result[idx] = resp[i].Object
-				}
-				return nil
-			})
-			if err := gr.Wait(); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	// concurrent repairs
-	gr, ctx := errgroup.WithContext(ctx)
-	for _, vote := range votes {
-		query := make([]*objects.VObject, 0, len(ids)/2)
-		for j, x := range lastTimes {
-			if cTime := vote.UpdateTimeAt(j); x.T != cTime && !x.Deleted {
-				query = append(query, &objects.VObject{LatestObject: &result[j].Object, StaleUpdateTime: cTime})
-			}
-		}
-		if len(query) == 0 {
-			continue
-		}
-		receiver := vote.Sender
-		gr.Go(func() error {
-			rs, err := cl.Overwrite(ctx, receiver, f.class, shard, query)
-			if err != nil {
-				return fmt.Errorf("node %q could not repair objects: %w", receiver, err)
-			}
-			for _, r := range rs {
-				if r.Err != "" {
-					return fmt.Errorf("object changed in the meantime on node %s: %s", receiver, r.Err)
-				}
-			}
-			return nil
-		})
-	}
-	err := gr.Wait()
-	if nDeletions > 0 {
-		return result, errConflictExistOrDeleted
-	}
-
-	return result, err
-}
-
-// iTuple tuple of indices used to identify a unique object
-type iTuple struct {
-	S       int   // sender's index
-	O       int   // object's index
-	T       int64 // last update time
-	Deleted bool
-}
-
-func (f *Finder) readExistence(ctx context.Context,
-	shard string,
-	id strfmt.UUID,
-	ch <-chan simpleResult[existReply],
-	st rState,
-) <-chan result[bool] {
-	resultCh := make(chan result[bool], 1)
-	go func() {
-		defer close(resultCh)
-		var (
-			votes    = make([]boolTuple, 0, len(st.Hosts)) // number of votes per replica
-			maxCount = 0
-		)
-
-		for r := range ch { // len(ch) == st.Level
-			resp := r.Response
-			if r.Err != nil { // a least one node is not responding
-				f.log.WithField("op", "exists").WithField("replica", resp.Sender).
-					WithField("class", f.class).WithField("shard", shard).
-					WithField("uuid", id).Error(r.Err)
-				resultCh <- result[bool]{false, errRead}
-				return
-			}
-
-			votes = append(votes, boolTuple{resp.Sender, resp.UpdateTime, resp.RepairResponse, 0, nil})
-			for i := range votes { // count number of votes
-				if votes[i].UTime == resp.UpdateTime {
-					votes[i].ack++
-				}
-				if maxCount < votes[i].ack {
-					maxCount = votes[i].ack
-				}
-				if maxCount >= st.Level {
-					exists := !votes[i].o.Deleted && votes[i].o.UpdateTime != 0
-					resultCh <- result[bool]{exists, nil}
-					return
-				}
-			}
-		}
-
-		obj, err := f.repairExist(ctx, shard, id, votes, st)
-		if err == nil {
-			resultCh <- result[bool]{obj, nil}
-			return
-		}
-		resultCh <- result[bool]{false, errRepair}
-
-		var sb strings.Builder
-		for i, c := range votes {
-			if i != 0 {
-				sb.WriteByte(' ')
-			}
-			fmt.Fprintf(&sb, "%s:%d", c.sender, c.UTime)
-		}
-		f.log.WithField("op", "repair_exist").WithField("class", f.class).
-			WithField("shard", shard).WithField("uuid", id).
-			WithField("msg", sb.String()).Error(err)
-	}()
-	return resultCh
-}
-
-func (f *Finder) repairExist(ctx context.Context, shard string, id strfmt.UUID, votes []boolTuple, st rState) (_ bool, err error) {
-	var (
-		lastUTime int64
-		winnerIdx int
-		cl        = f.client
-	)
-	for i, x := range votes {
-		if x.o.Deleted {
-			return false, errConflictExistOrDeleted
-		}
-		if x.UTime > lastUTime {
-			lastUTime = x.UTime
-			winnerIdx = i
-		}
-	}
-	// fetch most recent object
-	winner := votes[winnerIdx]
-	resp, err := cl.FullRead(ctx, winner.sender, f.class, shard, id, search.SelectProperties{}, additional.Properties{})
-	if err != nil {
-		return false, fmt.Errorf("get most recent object from %s: %w", winner.sender, err)
-	}
-	if resp.UpdateTime() != lastUTime {
-		return false, fmt.Errorf("fetch new state from %s: %w, %v", winner.sender, errConflictObjectChanged, err)
-	}
-	gr, ctx := errgroup.WithContext(ctx)
-	for _, vote := range votes { // repair
-		if vote.UTime == lastUTime {
-			continue
-		}
-		vote := vote
-		gr.Go(func() error {
-			ups := []*objects.VObject{{
-				LatestObject:    &resp.Object.Object,
-				StaleUpdateTime: vote.UTime,
-			}}
-			resp, err := cl.Overwrite(ctx, vote.sender, f.class, shard, ups)
-			if err != nil {
-				return fmt.Errorf("node %q could not repair object: %w", vote.sender, err)
-			}
-			if len(resp) > 0 && resp[0].Err != "" {
-				return fmt.Errorf("overwrite %w %s: %s", errConflictObjectChanged, vote.sender, resp[0].Err)
-			}
-			return nil
-		})
-	}
-	return !resp.Deleted, gr.Wait()
 }
