+++ usecases/replica/transport.go
@@ -27,6 +27,13 @@ const (
 	RequestKey = "request_id"
 )
 
+// Client is used to read and write objects on replicas
+type Client interface {
+	rClient
+	wClient
+}
+
+// StatusCode is communicate the cause of failure during replication
 type StatusCode int
 
 const (
@@ -65,7 +72,9 @@ func (e *Error) Clone() *Error {
 // Unwrap underlying error
 func (e *Error) Unwrap() error { return e.Err }
 
-func (e *Error) Error() string { return fmt.Sprintf("%s %q: %v", statusText(e.Code), e.Msg, e.Err) }
+func (e *Error) Error() string {
+	return fmt.Sprintf("%s %q: %v", statusText(e.Code), e.Msg, e.Err)
+}
 
 func (e *Error) IsStatusCode(sc StatusCode) bool {
 	return e.Code == sc
@@ -138,13 +147,24 @@ func (r *DeleteBatchResponse) FirstError() error {
 	return nil
 }
 
-// Client is used by replicator to communicate with other nodes
-type Client interface {
-	RClient
-	WClient
+type RepairResponse struct {
+	ID         string // object id
+	Version    int64  // sender's current version of the object
+	UpdateTime int64  // sender's current update time
+	Err        string
+	Deleted    bool
 }
 
-type WClient interface {
+func fromReplicas(xs []objects.Replica) []*storobj.Object {
+	rs := make([]*storobj.Object, len(xs))
+	for i := range xs {
+		rs[i] = xs[i].Object
+	}
+	return rs
+}
+
+// wClient is the client used to write to replicas
+type wClient interface {
 	PutObject(ctx context.Context, host, index, shard, requestID string,
 		obj *storobj.Object) (SimpleResponse, error)
 	DeleteObject(ctx context.Context, host, index, shard, requestID string,
@@ -161,8 +181,8 @@ type WClient interface {
 	Abort(ctx context.Context, host, index, shard, requestID string) (SimpleResponse, error)
 }
 
-// RClient is the client used to read from remote replicas
-type RClient interface {
+// rClient is the client used to read from remote replicas
+type rClient interface {
 	// FetchObject fetches one object
 	FetchObject(_ context.Context, host, index, shard string,
 		id strfmt.UUID, props search.SelectProperties,
@@ -184,37 +204,24 @@ type RClient interface {
 		ids []strfmt.UUID) ([]RepairResponse, error)
 }
 
-type RepairResponse struct {
-	ID         string // object id
-	Version    int64  // sender's current version of the object
-	UpdateTime int64  // sender's current update time
-	Err        string
-	Deleted    bool
-}
-
-func fromReplicas(xs []objects.Replica) []*storobj.Object {
-	rs := make([]*storobj.Object, len(xs))
-	for i := range xs {
-		rs[i] = xs[i].Object
-	}
-	return rs
-}
-
 // finderClient extends RClient with consistency checks
 type finderClient struct {
-	cl RClient
+	cl rClient
 }
 
 // FullRead reads full object
-func (fc finderClient) FullRead(ctx context.Context, host, index, shard string,
-	id strfmt.UUID, props search.SelectProperties,
+func (fc finderClient) FullRead(ctx context.Context,
+	host, index, shard string,
+	id strfmt.UUID,
+	props search.SelectProperties,
 	additional additional.Properties,
 ) (objects.Replica, error) {
 	return fc.cl.FetchObject(ctx, host, index, shard, id, props, additional)
 }
 
 // DigestReads reads digests of all specified objects
-func (fc finderClient) DigestReads(ctx context.Context, host, index, shard string,
+func (fc finderClient) DigestReads(ctx context.Context,
+	host, index, shard string,
 	ids []strfmt.UUID,
 ) ([]RepairResponse, error) {
 	n := len(ids)
@@ -226,7 +233,8 @@ func (fc finderClient) DigestReads(ctx context.Context, host, index, shard strin
 }
 
 // FullReads read full objects
-func (fc finderClient) FullReads(ctx context.Context, host, index, shard string,
+func (fc finderClient) FullReads(ctx context.Context,
+	host, index, shard string,
 	ids []strfmt.UUID,
 ) ([]objects.Replica, error) {
 	n := len(ids)
@@ -238,7 +246,8 @@ func (fc finderClient) FullReads(ctx context.Context, host, index, shard string,
 }
 
 // Overwrite specified object with most recent contents
-func (fc finderClient) Overwrite(ctx context.Context, host, index, shard string,
+func (fc finderClient) Overwrite(ctx context.Context,
+	host, index, shard string,
 	xs []*objects.VObject,
 ) ([]RepairResponse, error) {
 	return fc.cl.OverwriteObjects(ctx, host, index, shard, xs)
